renamed ExamView.js to AdminView.js
